Factored BaseTrunk into several files. converted degrees to radians in VectorRotation
